Refactor: move dynamic render field related code out of form
