+++ src/app/[locale]/incident/add/components/IncidentQuestionsLocationForm.tsx
@@ -4,17 +4,11 @@ import { useForm } from 'react-hook-form'
 import { useEffect, useState } from 'react'
 import { fetchAdditionalQuestions } from '@/services/additional-questions'
 import { useFormStore } from '@/store/form_store'
-import { LocationSelect } from '@/app/[locale]/incident/add/components/questions/LocationSelect'
 import { IncidentFormFooter } from '@/app/[locale]/incident/components/IncidentFormFooter'
 import { useStepperStore } from '@/store/stepper_store'
 import { useRouter } from '@/routing/navigation'
-import { RadioInput } from '@/app/[locale]/incident/add/components/questions/RadioInput'
-import { TextInput } from '@/app/[locale]/incident/add/components/questions/TextInput'
-import { PlainText } from '@/app/[locale]/incident/add/components/questions/PlainText'
-import { CheckboxInput } from '@/app/[locale]/incident/add/components/questions/CheckboxInput'
-import { TextAreaInput } from '@/app/[locale]/incident/add/components/questions/TextAreaInput'
-import { AssetSelect } from '@/app/[locale]/incident/add/components/questions/AssetSelect'
-import { FieldTypes, PublicQuestion } from '@/types/form'
+import { PublicQuestion } from '@/types/form'
+import { RenderDynamicFields } from '@/app/[locale]/incident/add/components/questions/RenderDynamicFields'
 
 export const IncidentQuestionsLocationForm = () => {
   const { formState: formStoreState, updateForm } = useFormStore()
@@ -34,33 +28,6 @@ export const IncidentQuestionsLocationForm = () => {
     router.prefetch('/incident/contact')
   }, [router])
 
-  // TODO: remove hardcoded marker
-  const marker = [0, 0]
-
-  const additionalQuestionTypes = {
-    [FieldTypes.RADIO_INPUT]: (field: PublicQuestion) => (
-      <RadioInput register={register} field={field} errors={errors} />
-    ),
-    [FieldTypes.PLAIN_TEXT]: (field: PublicQuestion) => (
-      <PlainText field={field} />
-    ),
-    [FieldTypes.TEXT_INPUT]: (field: PublicQuestion) => (
-      <TextInput register={register} field={field} errors={errors} />
-    ),
-    [FieldTypes.CHECKBOX_INPUT]: (field: PublicQuestion) => (
-      <CheckboxInput register={register} field={field} errors={errors} />
-    ),
-    [FieldTypes.TEXT_AREA_INPUT]: (field: PublicQuestion) => (
-      <TextAreaInput register={register} field={field} errors={errors} />
-    ),
-    [FieldTypes.ASSET_SELECT]: (field: PublicQuestion) => (
-      <AssetSelect field={field} />
-    ),
-    [FieldTypes.LOCATION_SELECT]: (props: any) => (
-      <LocationSelect marker={marker} {...props} />
-    ),
-  }
-
   useEffect(() => {
     const appendAdditionalQuestions = async () => {
       try {
@@ -68,10 +35,6 @@ export const IncidentQuestionsLocationForm = () => {
           formStoreState.main_category,
           formStoreState.sub_category
         )) as unknown as PublicQuestion[]
-
-        additionalQuestions.filter(
-          (question) => additionalQuestionTypes[question.field_type]
-        )
 
         setAdditionalQuestions(additionalQuestions)
         setLoading(false)
@@ -135,27 +98,17 @@ export const IncidentQuestionsLocationForm = () => {
     router.push('/incident/contact')
   }
 
-  const renderFields = (data: PublicQuestion[]) => {
-    return Object.keys(data).map((value, index, array) => {
-      const question = data[index]
-
-      const fieldName = question.key
-
-      return (
-        <div key={fieldName} className="w-full">
-          {additionalQuestionTypes[question.field_type]?.(question)}
-        </div>
-      )
-    })
-  }
-
   return (
     <form
       onSubmit={handleSubmit(onSubmit)}
       className="flex flex-col gap-8 items-start"
     >
       {additionalQuestions.length ? (
-        renderFields(additionalQuestions)
+        <RenderDynamicFields
+          data={additionalQuestions}
+          register={register}
+          errors={errors}
+        />
       ) : loading ? (
         /* TODO: Implement nice loading state */
         <p>Laden...</p>
